In p3, returns whole number

+++ march-27-2018/p3.py
@@ -1,7 +1,7 @@
 import math
 
 def compute(n):
-  return math.ceil(n)
+  return int(math.ceil(n))
   
 def start():
   N = input("Input: ")
